pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "agentgate"
version = "0.1.0"
description = "Runtime security for autonomous AI agents"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
keywords = ["ai", "agent", "security", "sandbox", "audit", "owasp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
crewai = ["crewai>=0.41"]
autogen = ["autogen-core>=0.4"]
all = ["agentgate[langchain,crewai,autogen]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
agentgate = "agentgate.cli.main:app"

[tool.maturin]
python-source = "python"
module-name = "agentgate._core"
features = ["pyo3/extension-module"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
